move refactor some guardian decryption items
